Implemented Logout
Signup, Login, and Logout are all working.

The navbar will now only display “Signup” and “Login” if there is
nobody logged in.
It will show the “Logout” button if the user is logged in. Upon
clicking this button, the user will be logged out.

*** TO DO ***
MAKE PAGE REFRESH ONCE USER LOGS IN, and LOGS OUT.